degugging:
Found out the intrsct.head->t of a cap intersection is being initialized as 0
and sometimes it's intrsct.head->next pointer is being NULL even if
intrsct.count is 2!

This second fact was causing a segfault. Checking for head->next NULL
instead of count solved the segfault but the color of the caps are all
being rendered as if all cap points were on t 0;