Found an issue but it's probably irrelevant:
Initialization of the .rows property of an inverted matrix is sometimes
getting lost in the stack. This property is only used to iterate the
rows when doing the inversion operations so it should be an issue only
if we use the inverse of an inverted matrix somewhere but we dont.

I'll continue reviewing the tests of the next chapters and if no other
issue show up, I'll get it fixed.